feat(App): loading each component as data arrives
partialy implementet
